Create Install SW Updates Available in Software Center on Devices
This script checks and installs the software updates available in Software Center on clients remotely with logging info, this is from a list within the script directory "Clients.txt". Log File goes to the Script Directory this is running from as Example: "c:\temp\ThisScriptFolder\InstallUpdates.log". Check ccmcache for SW Updates waiting to be be installed and installs them and WILL REBOOT THE SERVER ONCE COMPLETED!!!!!!